requirements.txt
requests>=2.25.1,<3.0.0
discord.py>=2.0.0,<3.0.0
psutil>=5.8.0,<6.0.0 